Refactor executable entry point
Move the executable implementation into a shared library entry module so
multiple binary shims can invoke it without changing behavior.
