+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod config;
 pub mod crate_command;
 pub mod dirs;
 pub mod discovery;
+pub mod entry;
 pub mod help_render;
 pub mod hook;
 pub mod hook_schema;
